fix(content-releases): refactor relation with entries
