+++ packages/core/content-releases/server/src/content-types/release-action/schema.ts
@@ -24,14 +24,12 @@ export default {
       enum: ['publish', 'unpublish'],
       required: true,
     },
-    entry: {
-      type: 'relation',
-      relation: 'morphToOne',
-      configurable: false,
-    },
     contentType: {
       type: 'string',
       required: true,
+    },
+    entryDocumentId: {
+      type: 'string',
     },
     locale: {
       type: 'string',
