+++ packages/core/content-releases/server/src/controllers/release.ts
@@ -13,7 +13,7 @@ import type {
   MapEntriesToReleases,
 } from '../../../shared/contracts/releases';
 import type { UserInfo } from '../../../shared/types';
-import { getService, getEntryId } from '../utils';
+import { getService } from '../utils';
 
 type ReleaseWithPopulatedActions = Release & { actions: { count: number } };
 
@@ -28,50 +28,38 @@ const releaseController = {
       ability: ctx.state.userAbility,
       model: RELEASE_MODEL_UID,
     });
-
     await permissionsManager.validateQuery(ctx.query);
-
     const releaseService = getService('release', { strapi });
-
     const query = await permissionsManager.sanitizeQuery(ctx.query);
-    const { contentTypeUid, documentId, hasEntryAttached, locale } = query;
-    const entryId = await getEntryId({ contentTypeUid, documentId, locale }, { strapi });
-
+    const { contentType, entryDocumentId, hasEntryAttached, locale } = query;
     const isEntryAttached =
       typeof hasEntryAttached === 'string' ? Boolean(JSON.parse(hasEntryAttached)) : false;
 
     if (isEntryAttached) {
       const releases = await releaseService.findMany({
         where: {
+          releasedAt: null,
           actions: {
-            target_type: contentTypeUid,
-            target_id: entryId,
-          },
-          releasedAt: {
-            $null: true,
+            contentType,
+            entryDocumentId: entryDocumentId ?? null,
+            locale: locale ?? null,
           },
         },
         populate: {
           actions: {
             fields: ['type'],
-            where: {
-              target_type: contentTypeUid,
-              target_id: entryId,
-            },
           },
         },
       });
-
       ctx.body = { data: releases };
     } else {
       const relatedReleases = await releaseService.findMany({
         where: {
-          releasedAt: {
-            $null: true,
-          },
+          releasedAt: null,
           actions: {
-            target_type: contentTypeUid,
-            target_id: entryId,
+            contentType,
+            entryDocumentId: entryDocumentId ?? null,
+            locale: locale ?? null,
           },
         },
       });
@@ -88,12 +76,9 @@ const releaseController = {
               actions: null,
             },
           ],
-          releasedAt: {
-            $null: true,
-          },
+          releasedAt: null,
         },
       });
-
       ctx.body = { data: releases };
     }
   },
@@ -137,12 +122,13 @@ const releaseController = {
     const id: GetRelease.Request['params']['id'] = ctx.params.id;
 
     const releaseService = getService('release', { strapi });
+    const releaseActionService = getService('release-action', { strapi });
     const release = await releaseService.findOne(id, { populate: ['createdBy'] });
     if (!release) {
       throw new errors.NotFoundError(`Release not found for id: ${id}`);
     }
 
-    const count = await releaseService.countActions({
+    const count = await releaseActionService.countActions({
       filters: {
         release: id,
       },
@@ -168,6 +154,7 @@ const releaseController = {
   },
 
   async mapEntriesToReleases(ctx: Koa.Context) {
+    /* @TODO: Migrate to new api 
     const { contentTypeUid, entriesIds } = ctx.query;
 
     if (!contentTypeUid || !entriesIds) {
@@ -176,7 +163,7 @@ const releaseController = {
 
     const releaseService = getService('release', { strapi });
 
-    const releasesWithActions = await releaseService.findManyWithContentTypeEntryAttached(
+    const releasesWithActions = await releaseService.findMany(
       contentTypeUid,
       entriesIds
     );
@@ -201,6 +188,7 @@ const releaseController = {
     ctx.body = {
       data: mappedEntriesInReleases,
     };
+    */
   },
 
   async create(ctx: Koa.Context) {
@@ -254,20 +242,20 @@ const releaseController = {
   },
 
   async publish(ctx: Koa.Context) {
-    const user: PublishRelease.Request['state']['user'] = ctx.state.user;
     const id: PublishRelease.Request['params']['id'] = ctx.params.id;
 
     const releaseService = getService('release', { strapi });
-    const release = await releaseService.publish(id, { user });
+    const releaseActionService = getService('release-action', { strapi });
+    const release = await releaseService.publish(id);
 
     const [countPublishActions, countUnpublishActions] = await Promise.all([
-      releaseService.countActions({
+      releaseActionService.countActions({
         filters: {
           release: id,
           type: 'publish',
         },
       }),
-      releaseService.countActions({
+      releaseActionService.countActions({
         filters: {
           release: id,
           type: 'unpublish',
