+++ packages/core/content-releases/server/src/services/validation.ts
@@ -24,7 +24,9 @@ const createReleaseValidationService = ({ strapi }: { strapi: Core.Strapi }) => 
       where: {
         id: releaseId,
       },
-      populate: { actions: { populate: { entry: { select: ['documentId', 'locale'] } } } },
+      populate: {
+        actions: true,
+      },
     })) as Release | null;
 
     if (!release) {
@@ -33,19 +35,20 @@ const createReleaseValidationService = ({ strapi }: { strapi: Core.Strapi }) => 
 
     const isEntryInRelease = release.actions.some(
       (action) =>
-        Number(action.entry.documentId) === Number(releaseActionArgs.entry.documentId) &&
-        action.contentType === releaseActionArgs.entry.contentType &&
-        action.locale === releaseActionArgs.entry.locale
+        Number(action.entryDocumentId) === Number(releaseActionArgs.entryDocumentId) &&
+        action.contentType === releaseActionArgs.contentType &&
+        action.locale === releaseActionArgs.locale
     );
 
     if (isEntryInRelease) {
       throw new AlreadyOnReleaseError(
-        `Entry with documentId ${releaseActionArgs.entry.documentId} ${releaseActionArgs.entry.locale ? `(${releaseActionArgs.entry.locale})` : ''} and contentType ${releaseActionArgs.entry.contentType} already exists in release with id ${releaseId}`
+        `Entry with documentId ${releaseActionArgs.entryDocumentId} ${releaseActionArgs.locale ? `(${releaseActionArgs.locale})` : ''} and contentType ${releaseActionArgs.contentType} already exists in release with id ${releaseId}`
       );
     }
   },
-  validateEntryContentType(
-    contentTypeUid: CreateReleaseAction.Request['body']['entry']['contentType']
+  validateEntryData(
+    contentTypeUid: CreateReleaseAction.Request['body']['contentType'],
+    entryDocumentId: CreateReleaseAction.Request['body']['entryDocumentId']
   ) {
     const contentType = strapi.contentType(contentTypeUid as UID.ContentType);
 
@@ -57,6 +60,10 @@ const createReleaseValidationService = ({ strapi }: { strapi: Core.Strapi }) => 
       throw new errors.ValidationError(
         `Content type with uid ${contentTypeUid} does not have draftAndPublish enabled`
       );
+    }
+
+    if (contentType.kind === 'collectionType' && !entryDocumentId) {
+      throw new errors.ValidationError('Document id is required for collection type');
     }
   },
   async validatePendingReleasesLimit() {
