+++ packages/core/content-releases/shared/contracts/release-actions.ts
@@ -15,6 +15,7 @@ export interface ReleaseAction extends Entity {
   type: 'publish' | 'unpublish';
   entry: ReleaseActionEntry;
   contentType: UID.ContentType;
+  entryDocumentId: ReleaseActionEntry['documentId'];
   locale?: string;
   release: Release;
   isEntryValid: boolean;
@@ -45,11 +46,9 @@ export declare namespace CreateReleaseAction {
     };
     body: {
       type: ReleaseAction['type'];
-      entry: {
-        documentId?: ReleaseActionEntry['documentId'];
-        locale?: ReleaseActionEntry['locale'];
-        contentType: string;
-      };
+      contentType: UID.ContentType;
+      entryDocumentId?: ReleaseActionEntry['documentId'];
+      locale?: ReleaseActionEntry['locale'];
     };
   }
